Update info.py
This command creates an embed, title has the users username, description has the users suggestion. If the user fails to input a suggestion the bot will detect this and remind them of it. The command that executes this is also deleted right away because there is no purpose in putting the suggestion in the suggestion channel if it sits in the bot command channel. 
This command also adds ✅ and ❎ to the suggestion so that users can easily vote.
Finally, in order to ensure anyone can make a suggestion, the suggestion command/syntax is listed in the footer, allowing users to see how to use the command.

--Jake.#8428 / JakeLion